fim restruturação security

+++ cloudKitchen/src/main/java/api5/cloudKitchen/security/Login.java
@@ -5,10 +5,10 @@ import lombok.Data;
 @Data
 public class Login {
 
-	private Integer idUsuario;
-	
+	private Long idUsuario;
+
     private String loginUsuario;
-    
+
     private String senhaUsuario;
 
     private String autorizacao;
